// src/pages.d.ts
interface EasyCom {
	/**
	 * 是否开启自动扫描，开启后将会自动扫描符合 `components/组件名称/组件名称.vue` 目录结构的组件
	 *
	 * @default true
	 */
	autoscan?: boolean;
	/**
     * 以正则方式自定义组件匹配规则
     *
     * 如果 `autoscan` 不能满足需求，可以使用 `custom` 自定义匹配规则
     *
     * @example
     * "custom": {
     *   "^uni-(.*)": "@/components/uni-$1.vue", // 匹配 components 目录内的 vue 文件
     *   "^vue-file-(.*)": "packageName/path/to/vue-file-$1.vue" // 匹配 node_modules 内的 vue 文件
     * }
  }
     */
	custom?: Record<string, string>;
}
// ...
export interface PageMetaDatum {
	/**
	 * 配置页面路径
	 */
	path: string;
	type?: string;
	/**
	 * 配置页面窗口表现，配置项参考下方 pageStyle
	 */
	style?: GlobalStyle;
	/**
	 * 当前页面是否需要登录才可以访问，此配置优先级高于 uniIdRouter 下的 needLogin
	 */
	needLogin?: boolean;
	[x: string]: any;
}
// ...
interface SubPackage {
	/**
	 * 子包的根目录
	 */
	root: string;
	/**
	 * 子包由哪些页面组成，参数同 pages
	 */
	pages: PageMetaDatum[];
	/**
	 * 分包插件
	 */
	plugins?: {
		[pluginName: string]: {
			version: string;
			provider: string;
			export?: string;
			[key: string]: any;
		};
	};
}
type SubPackages = SubPackage[];

interface TabBarIconFont {
	/**
	 * 字库 Unicode 码
	 */
	text?: string;
	/**
	 * 选中后字库 Unicode 码
	 */
	selectedText?: string;
	/**
	 * 字体图标字号，单位为 px
	 */
	fontSize?: string;
	/**
	 * 字体图标颜色
	 */
	color?: string;
	/**
	 * 字体图标选中颜色
	 */
	selectedColor?: string;
	[x: string]: any;
}
interface TabBarItem {
	/**
	 * 页面路径，必须在 pages 中先定义
	 */
	pagePath: string;
	/**
	 * tab 上按钮文字，在 App 和 H5 平台为非必填，例如中间可放一个没有文字的 + 号图标
	 */
	text?: string;
	/**
	 * 图片路径，icon 大小限制为 40 kb，建议尺寸为 81px * 81px
	 *
	 * 当 position 为 "top" 时，此参数无效
	 *
	 * 不支持网络图片，不支持字体图标
	 */
	iconPath?: string | string;
	/**
	 * 选中时的图片路径，icon 大小限制为 40 kb，建议尺寸为 81px * 81px
	 *
	 * 当 position 为 "top" 时，此参数无效
	 *
	 * 不支持网络图片，不支持字体图标
	 */
	selectedIconPath?: string | string;
	/**
	 * 该项是否显示，默认显示
	 *
	 * @desc App (3.2.10+)、H5 (3.2.10+)
	 */
	visible?: boolean;
	/**
	 * 字体图标，优先级高于 iconPath
	 *
	 * @desc App（3.4.4+）、H5 (3.5.3+)
	 */
	iconfont?: TabBarIconFont;
	[x: string]: any;
}
interface TabBarMidButton {
	/**
	 * 中间按钮的宽度，tabBar 其它项为减去此宽度后平分，默认值为与其它项平分宽度
	 *
	 * @default "80px"
	 */
	width?: string;
	/**
	 * 中间按钮的高度，可以大于 tabBar 高度，达到中间凸起的效果
	 *
	 * @default "50px"
	 */
	height?: string;
	/**
	 * 中间按钮的文字
	 */
	text?: string;
	/**
	 * 中间按钮的图片路径
	 */
	iconPath?: string;
	/**
	 * 图片宽度（高度等比例缩放）
	 *
	 * @default "24px"
	 */
	iconWidth?: string;
	/**
	 * 中间按钮的背景图片路径
	 */
	backgroundImage?: string;
	/**
	 * 字体图标，优先级高于 iconPath
	 *
	 * @desc App (3.4.4+)
	 */
	iconfont?: TabBarIconFont;
	[x: string]: any;
}
interface TabBar {
	/**
	 * tab 上的文字默认颜色
	 *
	 * @format color
	 */
	color?: string | string;
	/**
	 * tab 上的文字选中时的颜色
	 *
	 * @format color
	 */
	selectedColor?: string | string;
	/**
	 * tab 的背景色
	 *
	 * @format color
	 */
	backgroundColor?: string | string;
	/**
	 * tabBar 上边框的颜色
	 *
	 * @default "black"
	 *
	 * @desc App 2.3.4+、H5 3.0.0+
	 *
	 * @format color
	 */
	borderStyle?: "black" | "white" | string | string;
	/**
	 * iOS 高斯模糊效果，参考 [使用说明](https://ask.dcloud.net.cn/article/36617)
	 *
	 * @default "none"
	 *
	 * @desc App 2.4.0+、H5 3.0.0+（只有最新版浏览器才支持）
	 */
	blurEffect?: "dark" | "extralight" | "light" | "none";
	/**
	 * tab 列表，最少 2个，最多 5 个，app 端没有该限制
	 *
	 * @type {TabBarItem[]}
	 */
	list?: TabBarItem[];
	/**
	 * tab 位置
	 *
	 * @default "bottom"
	 *
	 * @desc "top" 仅微信小程序支持
	 */
	position?: "bottom" | "top";
	/**
	 * 文字默认大小
	 *
	 * @default "10px"
	 *
	 * @desc App 2.3.4+、H5 3.0.0+
	 */
	fontSize?: string;
	/**
	 * 图标默认宽度（高度等比例缩放）
	 *
	 * @default "24px"
	 *
	 * @desc App 2.3.4+、H5 3.0.0+
	 */
	iconWidth?: string;
	/**
	 * 图标和文字的间距
	 *
	 * @default "3px"
	 *
	 * @desc App 2.3.4+、H5 3.0.0+
	 */
	spacing?: string;
	/**
	 * tabBar 默认高度
	 *
	 * @default "50px"
	 *
	 * @desc App 2.3.4+、H5 3.0.0+
	 */
	height?: string;
	/**
	 * 中间按钮 仅在 list 项为偶数时有效
	 *
	 * @desc App 2.3.4+、H5 3.0.0+
	 */
	midButton?: TabBarMidButton;
	/**
	 * list 设置 iconfont 属性时，需要指定字体文件路径
	 *
	 * @desc App 3.4.4+、H5 3.5.3+
	 */
	iconfontSrc?: string;
	/**
	 * 设置背景图片，优先级高于 backgroundColor
	 */
	backgroundImage?: string;
	/**
	 * 设置标题栏的背景图平铺方式
	 *
	 * "repeat" 背景图片在垂直方向和水平方向平铺
	 *
	 * "repeat-x" 背景图片在水平方向平铺，垂直方向拉伸
	 *
	 * "repeat-y" 背景图片在垂直方向平铺，水平方向拉伸
	 *
	 * "no-repeat" 背景图片在垂直方向和水平方向都拉伸
	 *
	 * @default "no-repeat"
	 */
	backgroundRepeat?: "no-repeat" | "repeat" | "repeat-x" | "repeat-y";
	/**
	 * tabBar上红点颜色
	 *
	 * @format color
	 */
	redDotColor?: string;
	[x: string]: any;
}

interface TheWindow {
	/**
	 * 配置页面路径
	 */
	path: string;
	/**
	 * 配置页面窗口表现，配置项参考 [pageStyle](https://uniapp.dcloud.net.cn/collocation/pages#style)
	 */
	style?: GlobalStyle;
	/**
	 * 配置显示该窗口的规则，配置项参考 [matchMedia](https://uniapp.dcloud.net.cn/collocation/pages#matchmedia)
	 */
	matchMedia?: {
		/**
		 * 当设备可见区域宽度 >= minWidth 时，显示该 window
		 *
		 * @default 768
		 */
		minWidth?: number;
	};
}

interface UniIdRouter {
	/**
	 * 登录页面路径
	 *
	 * @example "pages/index/index"
	 */
	loginPage: string;
	/**
	 * 需要登录才可以访问的页面列表，可以使用正则语法
	 *
	 * @example ["pages/detail/.*"]
	 */
	needLogin: string[];
	/**
	 * 是否是否自动解析云对象及 clientDB 的错误码，如果是客户端 token 不正确或 token 过期则自动跳转配置的登录页面
	 *
	 * @default true
	 */
	resToLogin?: boolean;
}

/**
 * 对 uni-app 进行全局配置，决定页面文件的路径、窗口样式、原生的导航栏、底部的原生 tabBar 等，类似微信小程序中 app.json 的页面管理部分
 *
 * 注意定位权限申请等原属于 app.json 的内容，需要在 manifest 中配置
 */
export interface PagesConfig {
	/**
	 * 设置默认页面的窗口表现
	 */
	globalStyle?: GlobalStyle;
	/**
	 * 设置页面路径及窗口表现
	 */
	pages?: PageMetaDatum[];
	/**
	 * 组件自动引入规则
	 *
	 * @desc 2.5.5+
	 */
	easycom?: EasyCom;
	/**
	 * 设置底部 tab 的表现
	 */
	tabBar?: TabBar;

	/**
	 * 分包加载配置
	 *
	 * @desc H5 不支持
	 */
	subPackages?: SubPackages;
	/**
	 * 分包预下载规则
	 *
	 * @desc 微信小程序
	 */
	preloadRule?: {
		/**
		 * 页面路径
		 */
		[path: string]: {
			/**
			 * 进入页面后预下载分包的 root 或 name
			 *
			 * __APP__ 表示主包
			 */
			packages: string[];
			/**
			 * 在指定网络下预下载
			 *
			 * "all" 不限网络
			 *
			 * "wifi" 仅 wifi 下预下载
			 *
			 * @default "wifi"
			 */
			network?: "all" | "wifi";
		};
	};
	/**
	 * `Worker` 代码放置的目录
	 *
	 * @desc 微信小程序
	 */
	workers?:
		| string
		| {
				/**
				 * `Worker` 代码放置的目录
				 */
				path: string;
				/**
				 * 是否打包为分包
				 *
				 * @desc 微信小程序基础库 v2.27.3
				 *
				 * @default false
				 */
				isSubpackage?: boolean;
		  };
	/**
	 * 大屏左侧窗口
	 *
	 * @desc H5
	 */
	leftWindow?: TheWindow;
	/**
	 * 大屏顶部窗口
	 *
	 * @desc H5
	 */
	topWindow?: TheWindow;
	/**
	 * 大屏右侧窗口
	 *
	 * @desc H5
	 */
	rightWindow?: TheWindow;
	/**
	 * 自动跳转相关配置，新增于 HBuilderX 3.5.0
	 *
	 * @desc uni-app 3.5.0+、uni-app-x 3.99+
	 */
	uniIdRouter?: UniIdRouter;
	/**
	 * 默认启动首页，新增于 HBuilderX 3.7.0
	 *
	 * @desc 微信小程序、支付宝小程序
	 */
	entryPagePath?: string;
	[x: string]: any;
}